Remove old UpdateService
